Reorder to break up the dust.hpp file
