+++ inst/include/dust/dust.hpp
@@ -1,10 +1,6 @@
 #ifndef DUST_DUST_HPP
 #define DUST_DUST_HPP
 
-#include <dust/rng.hpp>
-#include <dust/densities.hpp>
-#include <dust/tools.hpp>
-
 #include <algorithm>
 #include <memory>
 #include <map>
@@ -15,409 +11,13 @@
 #include <omp.h>
 #endif
 
-
-// TODO: move these into a utilities file
-template <typename T, typename U, typename Enable = void>
-size_t destride_copy(T dest, U& src, size_t at, size_t stride) {
-  static_assert(!std::is_reference<T>::value,
-                "destride_copy should only be used with reference types");
-  size_t i;
-  for (i = 0; at < src.size(); ++i, at += stride) {
-    dest[i] = src[at];
-  }
-  return i;
-}
-
-template <typename T, typename U>
-size_t stride_copy(T dest, U src, size_t at, size_t stride) {
-  static_assert(!std::is_reference<T>::value,
-                "stride_copy should only be used with reference types");
-  dest[at] = src;
-  return at + stride;
-}
-
-template <typename T, typename U>
-size_t stride_copy(T dest, const std::vector<U>& src, size_t at, size_t stride) {
-  static_assert(!std::is_reference<T>::value,
-                "stride_copy should only be used with reference types");
-  for (size_t i = 0; i < src.size(); ++i, at += stride) {
-    dest[at] = src[i];
-  }
-  return at;
-}
-
-
-namespace dust {
-struct nothing {};
-typedef nothing no_data;
-typedef nothing no_internal;
-typedef nothing no_shared;
-
-// By default we do not support anything on the gpu. This name might
-// change, but it does reflect our intent and it's likely that to work
-// on a GPU the model will have to provide a number of things. If of
-// those becomes a type (as with data, internal and shared) we could
-// use the same approach as above.
-template <typename T>
-struct has_gpu_support : std::false_type {};
-
-template <typename T>
-using shared_ptr = std::shared_ptr<const typename T::shared_t>;
-
-template <typename T>
-struct pars_t {
-  std::shared_ptr<const typename T::shared_t> shared;
-  typename T::internal_t internal;
-
-  pars_t(std::shared_ptr<const typename T::shared_t> shared_,
-         typename T::internal_t internal_) :
-    shared(shared_), internal(internal_) {
-  }
-  pars_t(typename T::shared_t shared_,
-         typename T::internal_t internal_) :
-    shared(std::make_shared<const typename T::shared_t>(shared_)),
-    internal(internal_) {
-  }
-  pars_t(typename T::shared_t shared_) :
-    pars_t(shared_, dust::nothing()) {
-  }
-  pars_t(typename T::internal_t internal_) :
-    pars_t(dust::nothing(), internal_) {
-  }
-};
-
-template <typename real_t>
-class filter_state {
-public:
-  filter_state(size_t n_state, size_t n_particles, size_t n_data) :
-    n_state_(n_state), n_particles_(n_particles), n_data_(n_data), offset_(0) {
-    resize(n_state, n_particles, n_data);
-  }
-
-  // default constructable
-  filter_state() : filter_state(0, 0, 0) {
-  }
-
-  void resize(size_t n_state, size_t n_particles, size_t n_data) {
-    n_state_ = n_state;
-    n_particles_ = n_particles;
-    n_data_ = n_data;
-    offset_ = 0;
-    history_value.resize(n_state_ * n_particles_ * (n_data_ + 1));
-    history_order.resize(n_particles_ * (n_data_ + 1));
-    for (size_t i = 0; i < n_particles_; ++i) {
-      history_order[i] = i;
-    }
-  }
-
-  typename std::vector<real_t>::iterator history_value_iterator() {
-    return history_value.begin() + offset_ * n_state_ * n_particles_;
-  }
-
-  typename std::vector<size_t>::iterator history_order_iterator() {
-    return history_order.begin() + offset_ * n_particles_;
-  }
-
-  std::vector<real_t> history() const {
-    std::vector<real_t> ret(size());
-    history(ret.begin());
-    return ret;
-  }
-
-  // This is a particularly unpleasant bit of bookkeeping and is
-  // adapted from mcstate (see the helper files in tests for a
-  // translation of the the code). As we proceed we store the values
-  // of particles *before* resampling and then we store the index used
-  // in resampling. We do not resample all the history at each
-  // resample as that is prohibitively expensive.
-  //
-  // So to output sensible history we start with a particle and we
-  // look to see where it "came from" in the previous step
-  // (history_index) and propagate this backward in time to
-  // reconstruct what is in effect a multifurcating tree.
-  // This is analogous to the particle ancestor concept in the
-  // particle filter literature.
-  //
-  // It's possible we could do this more efficiently for some subset
-  // of particles too (give me the history of just one particle) by
-  // breaking the function before the loop over 'k'.
-  //
-  // Note that we treat history_order and history_value as read-only
-  // though this process so one could safely call this multiple times.
-  template <typename Iterator>
-  void history(Iterator ret) const {
-    std::vector<size_t> index_particle(n_particles_);
-    for (size_t i = 0; i < n_particles_; ++i) {
-      index_particle[i] = i;
-    }
-    for (size_t k = 0; k < n_data_ + 1; ++k) {
-      size_t i = n_data_ - k;
-      auto const it_order = history_order.begin() + i * n_particles_;
-      auto const it_value = history_value.begin() + i * n_state_ * n_particles_;
-      auto it_ret = ret + i * n_state_ * n_particles_;
-      for (size_t j = 0; j < n_particles_; ++j) {
-        const size_t idx = *(it_order + index_particle[j]);
-        index_particle[j] = idx;
-        std::copy_n(it_value + idx * n_state_, n_state_,
-                    it_ret + j * n_state_);
-      }
-    }
-  }
-
-  size_t size() const {
-    return history_value.size();
-  }
-
-  void advance() {
-    offset_++;
-  }
-
-private:
-  size_t n_state_;
-  size_t n_particles_;
-  size_t n_data_;
-  size_t offset_;
-  size_t len_;
-  std::vector<real_t> history_value;
-  std::vector<size_t> history_order;
-};
-
-template <typename real_t>
-struct device_state {
-  void initialise(size_t n_particles, size_t n_state, size_t n_shared_len_,
-                  size_t n_internal_int, size_t n_internal_real,
-                  size_t n_shared_int_, size_t n_shared_real_) {
-    n_shared_len = n_shared_len_;
-    n_shared_int = n_shared_int_;
-    n_shared_real = n_shared_real_;
-    const size_t n_rng = dust::rng_state_t<real_t>::size();
-    y = dust::device_array<real_t>(n_state * n_particles);
-    y_next = dust::device_array<real_t>(n_state * n_particles);
-    y_selected = dust::device_array<real_t>(n_state * n_particles);
-    internal_int = dust::device_array<int>(n_internal_int * n_particles);
-    internal_real = dust::device_array<real_t>(n_internal_real * n_particles);
-    shared_int = dust::device_array<int>(n_shared_int * n_shared_len);
-    shared_real = dust::device_array<real_t>(n_shared_real * n_shared_len);
-    rng = dust::device_array<uint64_t>(n_rng * n_particles);
-    index = dust::device_array<char>(n_state * n_particles);
-    n_selected = dust::device_array<int>(1);
-    scatter_index = dust::device_array<int>(n_state * n_particles);
-    set_cub_tmp();
-  }
-  void swap() {
-    std::swap(y, y_next);
-  }
-  // TODO - use GPU templates
-  void set_cub_tmp() {
-#ifdef __NVCC__
-    // Free the array before running cub function below
-    size_t tmp_bytes = 0;
-    select_tmp.set_size(tmp_bytes);
-    cub::DeviceSelect::Flagged(select_tmp.data(),
-                               tmp_bytes,
-                               y.data(),
-                               index.data(),
-                               y_selected.data(),
-                               n_selected.data(),
-                               y.size());
-    select_tmp.set_size(tmp_bytes);
-#endif
-  }
-
-  size_t n_shared_len;
-  size_t n_shared_int;
-  size_t n_shared_real;
-  dust::device_array<real_t> y;
-  dust::device_array<real_t> y_next;
-  dust::device_array<real_t> y_selected;
-  dust::device_array<int> internal_int;
-  dust::device_array<real_t> internal_real;
-  dust::device_array<int> shared_int;
-  dust::device_array<real_t> shared_real;
-  dust::device_array<uint64_t> rng;
-  dust::device_array<char> index;
-  dust::device_array<int> n_selected;
-  dust::device_array<void> select_tmp;
-  dust::device_array<int> scatter_index;
-};
-
-// We need to compute the size of space required for integers and
-// reals on the device, per particle. Because we work on the
-// requirement that every particle has the same dimension we pass an
-// arbitrary set of shared parameters (really the first) to
-// device_internal_size. The underlying model can overload this template
-// for either real or int types and return the length of data
-// required.
-template <typename T>
-size_t device_internal_size_int(typename dust::shared_ptr<T> shared) {
-  return 0;
-}
-
-template <typename T>
-size_t device_internal_size_real(typename dust::shared_ptr<T> shared) {
-  return 0;
-}
-
-template <typename T>
-size_t device_shared_size_int(typename dust::shared_ptr<T> shared) {
-  return 0;
-}
-
-template <typename T>
-size_t device_shared_size_real(typename dust::shared_ptr<T> shared) {
-  return 0;
-}
-
-template <typename T>
-void device_shared_copy(typename dust::shared_ptr<T> shared,
-                        int * shared_int,
-                        typename T::real_t * shared_real) {
-}
-
-template <typename T>
-T* shared_copy(T* dest, const std::vector<T>& src) {
-  memcpy(dest, src.data(), src.size() * sizeof(T));
-  return dest + src.size();
-}
-
-template <typename T>
-T* shared_copy(T* dest, const T src) {
-  *dest = src;
-  return dest + 1;
-}
-
-}
-
-// We'll need to expand this soon to cope with shared memory, but that
-// will be coming via another change to dust. Or we can do it here
-// with a shared object that contains just:
-// template <typename T>
-// struct shared_t {
-//   const real_t const * real_data;
-//   const int * const * int_data;
-//   // plus lengths?
-// };
-//
-// which can just point at the data in the vector because it will live
-// longer than the object.
-template <typename T>
-DEVICE void update_device(size_t step,
-                   const dust::interleaved<typename T::real_t> state,
-                   dust::interleaved<int> internal_int,
-                   dust::interleaved<typename T::real_t> internal_real,
-                   const int * shared_int,
-                   const typename T::real_t * shared_real,
-                   dust::rng_state_t<typename T::real_t>& rng_state,
-                   dust::interleaved<typename T::real_t> state_next);
-
-// __global__ for shuffling particles
-template<typename real_t>
-KERNEL void device_scatter(int* scatter_index,
-                           real_t* state,
-                           real_t* scatter_state,
-                           size_t state_size) {
-#ifdef __NVCC__
-  int i = blockIdx.x * blockDim.x + threadIdx.x;
-  if (i < state_size) {
-#else
-  for (size_t i = 0; i < state_size; ++i) {
-#endif
-    scatter_state[i] = state[scatter_index[i]];
-  }
-}
-
-template <typename T>
-KERNEL void run_particles(size_t step_start, size_t step_end, size_t n_particles,
-                   size_t n_pars,
-                   typename T::real_t * state, typename T::real_t * state_next,
-                   int * internal_int, typename T::real_t * internal_real,
-                   size_t n_shared_int, size_t n_shared_real,
-                   const int * shared_int,
-                   const typename T::real_t * shared_real,
-                   uint64_t * rng_state,
-                   bool use_shared_L1);
-
-template <typename T>
-class Particle {
-public:
-  typedef dust::pars_t<T> pars_t;
-  typedef typename T::real_t real_t;
-  typedef typename T::data_t data_t;
-
-  Particle(pars_t pars, size_t step) :
-    _model(pars),
-    _step(step),
-    _y(_model.initial(_step)),
-    _y_swap(_model.size()) {
-  }
-
-  void run(const size_t step_end, dust::rng_state_t<real_t>& rng_state) {
-    while (_step < step_end) {
-      _model.update(_step, _y.data(), rng_state, _y_swap.data());
-      _step++;
-      std::swap(_y, _y_swap);
-    }
-  }
-
-  void state(const std::vector<size_t>& index,
-             typename std::vector<real_t>::iterator end_state) const {
-    for (size_t i = 0; i < index.size(); ++i) {
-      *(end_state + i) = _y[index[i]];
-    }
-  }
-
-  void state_full(typename std::vector<real_t>::iterator end_state) const {
-    for (size_t i = 0; i < _y.size(); ++i) {
-      *(end_state + i) = _y[i];
-    }
-  }
-
-  size_t size() const {
-    return _y.size();
-  }
-
-  size_t step() const {
-    return _step;
-  }
-
-  void swap() {
-    std::swap(_y, _y_swap);
-  }
-
-  void set_step(const size_t step) {
-    _step = step;
-  }
-
-  void set_state(const Particle<T>& other) {
-    _y_swap = other._y;
-  }
-
-  void set_pars(const Particle<T>& other, bool set_state) {
-    _model = other._model;
-    _step = other._step;
-    if (set_state) {
-      _y = _model.initial(_step);
-    }
-  }
-
-  void set_state(typename std::vector<real_t>::const_iterator state) {
-    for (size_t i = 0; i < _y.size(); ++i, ++state) {
-      _y[i] = *state;
-    }
-  }
-
-  real_t compare_data(const data_t& data,
-                      dust::rng_state_t<real_t>& rng_state) {
-    return _model.compare_data(_y.data(), data, rng_state);
-  }
-
-private:
-  T _model;
-  size_t _step;
-
-  std::vector<real_t> _y;
-  std::vector<real_t> _y_swap;
-};
+#include <dust/rng.hpp>
+#include <dust/densities.hpp>
+#include <dust/tools.hpp>
+#include <dust/types.hpp>
+#include <dust/utils.hpp>
+#include <dust/particle.hpp>
+#include <dust/kernels.hpp>
 
 template <typename T>
 class Dust {
@@ -664,7 +264,8 @@ public:
       #pragma omp parallel for schedule(static) num_threads(_n_threads)
 #endif
       for (size_t i = 0; i < np; ++i) {
-        destride_copy(end_state + i * index_size, y_selected, i, np);
+        dust::utils::destride_copy(end_state + i * index_size, y_selected, i,
+                                   np);
       }
 #else
       refresh_host();
@@ -743,14 +344,14 @@ public:
 #ifdef __NVCC__
       const size_t blockSize = 128;
       const size_t blockCount = (scatter_state.size() + blockSize - 1) / blockSize;
-      device_scatter<real_t><<<blockCount, blockSize>>>(
+      scatter_device<real_t><<<blockCount, blockSize>>>(
         _device_data.scatter_index.data(),
         _device_data.y.data(),
         _device_data.y_next.data(),
         scatter_state.size());
       CUDA_CALL(cudaDeviceSynchronize());
 #else
-      device_scatter<real_t>(
+      scatter_device<real_t>(
         _device_data.scatter_index.data(),
         _device_data.y.data(),
         _device_data.y_next.data(),
@@ -1166,13 +767,14 @@ private:
       for (size_t i = 0; i < np; ++i) {
         // Interleave state
         _particles[i].state_full(y_tmp.begin());
-        stride_copy(y.data(), y_tmp, i, np);
+        dust::utils::stride_copy(y.data(), y_tmp, i, np);
 
         // Interleave RNG state
         dust::rng_state_t<real_t> p_rng = _rng.state(i);
         size_t rng_offset = i;
         for (size_t j = 0; j < rng_len; ++j) {
-          rng_offset = stride_copy(rng.data(), p_rng[j], rng_offset, np);
+          rng_offset = dust::utils::stride_copy(rng.data(), p_rng[j],
+                                                rng_offset, np);
         }
       }
       // H -> D copies
@@ -1199,7 +801,7 @@ private:
       #pragma omp parallel for schedule(static) num_threads(_n_threads)
 #endif
       for (size_t i = 0; i < np; ++i) {
-        destride_copy(y_tmp.data(), y, i, np);
+        dust::utils::destride_copy(y_tmp.data(), y, i, np);
         _particles[i].set_state(y_tmp.begin());
 
         // Destride RNG
@@ -1212,93 +814,5 @@ private:
     }
   }
 };
-
-template <typename T>
-KERNEL void run_particles(size_t step_start, size_t step_end, size_t n_particles,
-                   size_t n_pars,
-                   typename T::real_t * state, typename T::real_t * state_next,
-                   int * internal_int, typename T::real_t * internal_real,
-                   size_t n_shared_int, size_t n_shared_real,
-                   const int * shared_int,
-                   const typename T::real_t * shared_real,
-                   uint64_t * rng_state,
-                   bool use_shared_L1) {
-  typedef typename T::real_t real_t;
-  const size_t n_particles_each = n_particles / n_pars;
-
-#ifdef __CUDA_ARCH__
-  // Particle index i, and max index to process in the block
-  int i, max_i;
-
-  // Get pars index j, and start address in shared space
-  const int block_per_pars = (n_particles_each + blockDim.x - 1) / blockDim.x;
-  const int j = blockIdx.x / block_per_pars;
-  const int * p_shared_int = shared_int + j * n_shared_int;
-  const real_t * p_shared_real = shared_real + j * n_shared_real;
-
-  // If we're using it, use the first warp in the block to load the shared pars
-  // into __shared__ L1
-  extern __shared__ int shared_block[];
-  auto block = cooperative_groups::this_thread_block();
-  if (use_shared_L1) {
-    int * shared_block_int = shared_block;
-    shared_mem_cpy(block, shared_block_int, p_shared_int, n_shared_int);
-    p_shared_int = shared_block_int;
-
-    // Must only have a single __shared__ definition, cast to use different
-    // types within it
-    real_t * shared_block_real = (real_t*)&shared_block[n_shared_int];
-    shared_mem_cpy(block, shared_block_real, p_shared_real, n_shared_real);
-    p_shared_real = shared_block_real;
-
-    // Pick particle index based on block, don't process if off the end
-    i = j * n_particles_each + (blockIdx.x % block_per_pars) * blockDim.x + threadIdx.x;
-    max_i = n_particles_each * (j + 1);
-  } else {
-    // Otherwise CUDA thread number = particle
-    i = blockIdx.x * blockDim.x + threadIdx.x;
-    max_i = n_particles;
-  }
-
-  // Required to sync loads into L1 cache
-  shared_mem_wait(block);
-
-  if (i < max_i) {
-#else
-  // omp here
-  for (size_t i = 0; i < n_particles; ++i) {
-    const int j = i / n_particles_each;
-    const int * p_shared_int = shared_int + j * n_shared_int;
-    const real_t * p_shared_real = shared_real + j * n_shared_real;
-#endif
-    dust::interleaved<real_t> p_state(state, i, n_particles);
-    dust::interleaved<real_t> p_state_next(state_next, i, n_particles);
-    dust::interleaved<int> p_internal_int(internal_int, i, n_particles);
-    dust::interleaved<real_t> p_internal_real(internal_real, i, n_particles);
-    dust::interleaved<uint64_t> p_rng(rng_state, i, n_particles);
-
-    dust::rng_state_t<real_t> rng_block = dust::get_rng_state<real_t>(p_rng);
-    for (size_t step = step_start; step < step_end; ++step) {
-      update_device<T>(step,
-                       p_state,
-                       p_internal_int,
-                       p_internal_real,
-                       p_shared_int,
-                       p_shared_real,
-                       rng_block,
-                       p_state_next);
-#ifdef __CUDA_ARCH__
-      __syncwarp();
-#endif
-      dust::interleaved<real_t> tmp = p_state;
-      p_state = p_state_next;
-      p_state_next = tmp;
-    }
-    dust::put_rng_state(rng_block, p_rng);
-#ifdef __CUDA_ARCH__
-    block.sync();
-#endif
-  }
-}
 
 #endif
