+++ inst/include/dust/particle.hpp
@@ -1,0 +1,86 @@
+#ifndef DUST_PARTICLE_HPP
+#define DUST_PARTICLE_HPP
+
+template <typename T>
+class Particle {
+public:
+  typedef dust::pars_t<T> pars_t;
+  typedef typename T::real_t real_t;
+  typedef typename T::data_t data_t;
+
+  Particle(pars_t pars, size_t step) :
+    _model(pars),
+    _step(step),
+    _y(_model.initial(_step)),
+    _y_swap(_model.size()) {
+  }
+
+  void run(const size_t step_end, dust::rng_state_t<real_t>& rng_state) {
+    while (_step < step_end) {
+      _model.update(_step, _y.data(), rng_state, _y_swap.data());
+      _step++;
+      std::swap(_y, _y_swap);
+    }
+  }
+
+  void state(const std::vector<size_t>& index,
+             typename std::vector<real_t>::iterator end_state) const {
+    for (size_t i = 0; i < index.size(); ++i) {
+      *(end_state + i) = _y[index[i]];
+    }
+  }
+
+  void state_full(typename std::vector<real_t>::iterator end_state) const {
+    for (size_t i = 0; i < _y.size(); ++i) {
+      *(end_state + i) = _y[i];
+    }
+  }
+
+  size_t size() const {
+    return _y.size();
+  }
+
+  size_t step() const {
+    return _step;
+  }
+
+  void swap() {
+    std::swap(_y, _y_swap);
+  }
+
+  void set_step(const size_t step) {
+    _step = step;
+  }
+
+  void set_state(const Particle<T>& other) {
+    _y_swap = other._y;
+  }
+
+  void set_pars(const Particle<T>& other, bool set_state) {
+    _model = other._model;
+    _step = other._step;
+    if (set_state) {
+      _y = _model.initial(_step);
+    }
+  }
+
+  void set_state(typename std::vector<real_t>::const_iterator state) {
+    for (size_t i = 0; i < _y.size(); ++i, ++state) {
+      _y[i] = *state;
+    }
+  }
+
+  real_t compare_data(const data_t& data,
+                      dust::rng_state_t<real_t>& rng_state) {
+    return _model.compare_data(_y.data(), data, rng_state);
+  }
+
+private:
+  T _model;
+  size_t _step;
+
+  std::vector<real_t> _y;
+  std::vector<real_t> _y_swap;
+};
+
+#endif
